prob5 compiles but is broken

+++ prob5/src/primes.rs
@@ -4,8 +4,8 @@
 //   1. Cross out multiples of 2, then multiples of 3, etc...
 //   2. By "cross out" we probably mean flip a bit in `mask`
 
-pub fn primes(n: u64) -> Vec<u64> {
-    let nums: Vec<u64> = (2..=(n+3)).collect();
+pub fn primes(n: i32) -> Vec<i32> {
+    let nums: Vec<i32> = (2..=(n+3)).collect();
     let mut not_prime = vec![false; (n+1) as usize];
     
     for i in 0..=n as usize {
@@ -31,7 +31,7 @@ pub fn primes(n: u64) -> Vec<u64> {
             }   
         }
     }
-    let mut prime_nums: Vec<u64> = Vec::new();
+    let mut prime_nums: Vec<i32> = Vec::new();
     for i in 0..=n as usize {
         if not_prime[i] {
             continue;
